adjusted creepage distance
The closest inner layer to outer layer, which uses a prepreg of 0.21 mm, would result in 0.81 mm creepage distance on edges. Due to unknown IPC 2221 rating for this case, the design was adjusted (below the gate driver and a trace below the resonant capacitors) from 0.3 mm to 0.6 mm to be safe.