Added R75 to lower the fan turn-on temperature
Note: R75 (100k) is not added in the PCB design due to space restrictions. It is paralleled to R31 (4.7k) so the parallel value is around 4.49 k. The turn-on temperature is now 72°C and turn-off 58°C.
The original design was measured to have 71°C/61°C on/off. However it was found that when mounted, the transformer could reach 90°C on the top without turning on the fan. A new measurement showed, that it actually turned on at 84°C and 70°C off, but everything else (e.g. voltages) seem to be the same from the initial measurement. There might be a bias due to the initial value being measured without being mounted to the transformer. Anyways, the lowered threshold will prolong the component's lifetime due to lower thermal stress. Soldering R75 on top of R31 is recommended but not 100% needed.
Replacing R31 with 4.3 k is possible too, but the fan will probably run all the time due to an even lower turn on/off threshold.