IMPORTANT NOTICE see #10 and fix in iBom
IMPORTANT NOTICE: There was a complaint about the converter not being rules conform. Due to other personal priorities it might take some weeks to solve the addressed points. There is also the chance of some points unable to be resolved. See #10 for further details.
iBom: There were some components missing in the iBom for PCBA.